+++ CHANGELOG.md
@@ -8,9 +8,10 @@
   features currently remain exclusive to unsigned integers:
     * Support for the following optional Cargo features: `step_trait`, `borsh` and `schemars`
     * Byte operations (`to_ne_bytes`, ...)
-    * Extract functions (`extract_i8`, ...)
     * Checked arithmetic functions (`checked_add`, ...)
     * Overflowing arithmetic functions (`overflowing_add`, ...)
+- Various new extract functions: `extract_i8`, `extract_i16`, ..., `extract_i128`. These are the same as the
+  equivalent `extract_u<N>` functions, but work with signed integers instead.
 
 ### Fixed